Fix handling of idxEnv in sequential mode code gen of 1-dim scans, create default code gen functions
In the sequential mode of the code generation, as used by 1-dimensional scans, we used loops instead of loops', which meant that the list of loops also contained the outermost loop. That outermost loop shouldn't be in this list as it was handled separately.

This commit also introduces default implementations of simple operations, like generate and map. These can be used by most backends. In the future, we may extend this with more default implementations.